refactor and remove dead code
