+++ validator/db/iface/interface.go
@@ -5,8 +5,6 @@ import (
 	"context"
 	"io"
 
-	"github.com/prysmaticlabs/go-bitfield"
-	slashpb "github.com/prysmaticlabs/prysm/proto/slashing"
 	"github.com/prysmaticlabs/prysm/validator/db/kv"
 )
 
@@ -21,24 +19,16 @@ type ValidatorDB interface {
 	GenesisValidatorsRoot(ctx context.Context) ([]byte, error)
 	SaveGenesisValidatorsRoot(ctx context.Context, genValRoot []byte) error
 
-	// Proposer protection related methods.
-	ProposalHistoryForEpoch(ctx context.Context, publicKey []byte, epoch uint64) (bitfield.Bitlist, error)
-	SaveProposalHistoryForEpoch(ctx context.Context, publicKey []byte, epoch uint64, history bitfield.Bitlist) error
-
-	// New data structure methods
+	// Proposer history methods.
 	HighestSignedProposal(ctx context.Context, publicKey [48]byte) (uint64, error)
 	LowestSignedProposal(ctx context.Context, publicKey [48]byte) (uint64, error)
 	ProposalHistoryForSlot(ctx context.Context, publicKey [48]byte, slot uint64) ([]byte, error)
 	SaveProposalHistoryForSlot(ctx context.Context, pubKey [48]byte, slot uint64, signingRoot []byte) error
 	ProposedPublicKeys(ctx context.Context) ([][48]byte, error)
 
-	// Attester protection related methods.
-	AttestationHistoryForPubKeys(ctx context.Context, publicKeys [][48]byte) (map[[48]byte]*slashpb.AttestationHistory, error)
-	SaveAttestationHistoryForPubKeys(ctx context.Context, historyByPubKey map[[48]byte]*slashpb.AttestationHistory) error
-
-	// New attestation store methods.
-	AttestationHistoryForPubKeysV2(ctx context.Context, publicKeys [][48]byte) (map[[48]byte]kv.EncHistoryData, error)
-	SaveAttestationHistoryForPubKeysV2(ctx context.Context, historyByPubKeys map[[48]byte]kv.EncHistoryData) error
-	SaveAttestationHistoryForPubKeyV2(ctx context.Context, pubKey [48]byte, history kv.EncHistoryData) error
+	// Attester history methods.
+	AttestationHistoryForPubKeys(ctx context.Context, publicKeys [][48]byte) (map[[48]byte]kv.EncHistoryData, error)
+	SaveAttestationHistoryForPubKeys(ctx context.Context, historyByPubKeys map[[48]byte]kv.EncHistoryData) error
+	SaveAttestationHistoryForPubKey(ctx context.Context, pubKey [48]byte, history kv.EncHistoryData) error
 	AttestedPublicKeys(ctx context.Context) ([][48]byte, error)
 }
