+++ validator/db/kv/db_test.go
@@ -10,7 +10,7 @@ import (
 func setupDB(t testing.TB, pubkeys [][48]byte) *Store {
 	db, err := NewKVStore(t.TempDir(), pubkeys)
 	require.NoError(t, err, "Failed to instantiate DB")
-	err = db.OldUpdatePublicKeysBuckets(pubkeys)
+	err = db.UpdatePublicKeysBuckets(pubkeys)
 	require.NoError(t, err, "Failed to create old buckets for public keys")
 	t.Cleanup(func() {
 		require.NoError(t, db.Close(), "Failed to close database")
