+++ validator/db/kv/proposal_history.go
@@ -2,13 +2,12 @@ package kv
 
 import (
 	"context"
-	"encoding/binary"
 	"fmt"
 
 	"github.com/pkg/errors"
-	"github.com/prysmaticlabs/go-bitfield"
+	"github.com/prysmaticlabs/prysm/beacon-chain/core/helpers"
+	"github.com/prysmaticlabs/prysm/shared/bytesutil"
 	"github.com/prysmaticlabs/prysm/shared/params"
-	"github.com/wealdtech/go-bytesutil"
 	bolt "go.etcd.io/bbolt"
 	"go.opencensus.io/trace"
 )
@@ -23,55 +22,146 @@ type Proposal struct {
 	SigningRoot []byte `json:"signing_root"`
 }
 
-// ProposalHistoryForEpoch accepts a validator public key and returns the corresponding proposal history.
-// Returns nil if there is no proposal history for the validator.
-func (store *Store) ProposalHistoryForEpoch(ctx context.Context, publicKey []byte, epoch uint64) (bitfield.Bitlist, error) {
-	ctx, span := trace.StartSpan(ctx, "Validator.ProposalHistoryForEpoch")
+// ProposedPublicKeys retrieves all public keys in our proposals history bucket.
+func (store *Store) ProposedPublicKeys(ctx context.Context) ([][48]byte, error) {
+	ctx, span := trace.StartSpan(ctx, "Validator.ProposedPublicKeys")
+	defer span.End()
+	var err error
+	proposedPublicKeys := make([][48]byte, 0)
+	err = store.view(func(tx *bolt.Tx) error {
+		bucket := tx.Bucket(newhistoricProposalsBucket)
+		return bucket.ForEach(func(key []byte, _ []byte) error {
+			pubKeyBytes := [48]byte{}
+			copy(pubKeyBytes[:], key)
+			proposedPublicKeys = append(proposedPublicKeys, pubKeyBytes)
+			return nil
+		})
+	})
+	return proposedPublicKeys, err
+}
+
+// ProposalHistoryForSlot accepts a validator public key and returns the corresponding signing root.
+// Returns nil if there is no proposal history for the validator at this slot.
+func (store *Store) ProposalHistoryForSlot(ctx context.Context, publicKey [48]byte, slot uint64) ([]byte, error) {
+	ctx, span := trace.StartSpan(ctx, "Validator.ProposalHistoryForSlot")
 	defer span.End()
 
 	var err error
-	// Adding an extra byte for the bitlist length.
-	slotBitlist := make(bitfield.Bitlist, params.BeaconConfig().SlotsPerEpoch/8+1)
+	signingRoot := make([]byte, 32)
 	err = store.view(func(tx *bolt.Tx) error {
-		bucket := tx.Bucket(historicProposalsBucket)
-		valBucket := bucket.Bucket(publicKey)
+		bucket := tx.Bucket(newhistoricProposalsBucket)
+		valBucket := bucket.Bucket(publicKey[:])
 		if valBucket == nil {
-			return fmt.Errorf("validator history empty for public key %#x", publicKey)
+			return fmt.Errorf("validator history empty for public key: %#x", publicKey)
 		}
-		slotBits := valBucket.Get(bytesutil.Bytes8(epoch))
-		if len(slotBits) == 0 {
-			slotBitlist = bitfield.NewBitlist(params.BeaconConfig().SlotsPerEpoch)
+		sr := valBucket.Get(bytesutil.Uint64ToBytesBigEndian(slot))
+		if len(sr) == 0 {
 			return nil
 		}
-		copy(slotBitlist, slotBits)
+		copy(signingRoot, sr)
 		return nil
 	})
-	return slotBitlist, err
+	return signingRoot, err
 }
 
-// SaveProposalHistoryForEpoch saves the proposal history for the requested validator public key.
-func (store *Store) SaveProposalHistoryForEpoch(ctx context.Context, pubKey []byte, epoch uint64, slotBits bitfield.Bitlist) error {
+// SaveProposalHistoryForSlot saves the proposal history for the requested validator public key.
+// We also check if the incoming proposal slot is lower than the lowest signed proposal slot
+// for the validator and override its value on disk.
+func (store *Store) SaveProposalHistoryForSlot(ctx context.Context, pubKey [48]byte, slot uint64, signingRoot []byte) error {
 	ctx, span := trace.StartSpan(ctx, "Validator.SaveProposalHistoryForEpoch")
 	defer span.End()
 
 	err := store.update(func(tx *bolt.Tx) error {
-		bucket := tx.Bucket(historicProposalsBucket)
-		valBucket := bucket.Bucket(pubKey)
-		if valBucket == nil {
-			return fmt.Errorf("validator history is empty for validator %#x", pubKey)
+		bucket := tx.Bucket(newhistoricProposalsBucket)
+		valBucket, err := bucket.CreateBucketIfNotExists(pubKey[:])
+		if err != nil {
+			return fmt.Errorf("could not create bucket for public key %#x", pubKey)
 		}
-		if err := valBucket.Put(bytesutil.Bytes8(epoch), slotBits); err != nil {
+
+		// If the incoming slot is lower than the lowest signed proposal slot, override.
+		lowestSignedProposalBytes := valBucket.Get(lowestSignedProposalKey)
+		var lowestSignedProposalSlot uint64
+		if len(lowestSignedProposalBytes) != 0 {
+			lowestSignedProposalSlot = bytesutil.BytesToUint64BigEndian(lowestSignedProposalBytes)
+		}
+		if len(lowestSignedProposalBytes) == 0 || slot < lowestSignedProposalSlot {
+			if err := valBucket.Put(lowestSignedProposalKey, bytesutil.Uint64ToBytesBigEndian(slot)); err != nil {
+				return err
+			}
+		}
+
+		// If the incoming slot is higher than the highest signed proposal slot, override.
+		highestSignedProposalBytes := valBucket.Get(highestSignedProposalKey)
+		var highestSignedProposalSlot uint64
+		if len(highestSignedProposalBytes) != 0 {
+			highestSignedProposalSlot = bytesutil.BytesToUint64BigEndian(highestSignedProposalBytes)
+		}
+		if len(highestSignedProposalBytes) == 0 || slot > highestSignedProposalSlot {
+			if err := valBucket.Put(highestSignedProposalKey, bytesutil.Uint64ToBytesBigEndian(slot)); err != nil {
+				return err
+			}
+		}
+
+		if err := valBucket.Put(bytesutil.Uint64ToBytesBigEndian(slot), signingRoot); err != nil {
 			return err
 		}
-		return pruneProposalHistory(valBucket, epoch)
+		return pruneProposalHistoryBySlot(valBucket, slot)
 	})
 	return err
 }
 
+// LowestSignedProposal returns the lowest signed proposal slot for a validator public key.
+// If no data exists, returning 0 is a sensible default.
+func (store *Store) LowestSignedProposal(ctx context.Context, publicKey [48]byte) (uint64, error) {
+	ctx, span := trace.StartSpan(ctx, "Validator.LowestSignedProposal")
+	defer span.End()
+
+	var err error
+	var lowestSignedProposalSlot uint64
+	err = store.view(func(tx *bolt.Tx) error {
+		bucket := tx.Bucket(newhistoricProposalsBucket)
+		valBucket := bucket.Bucket(publicKey[:])
+		if valBucket == nil {
+			return fmt.Errorf("validator history empty for public key: %#x", publicKey)
+		}
+		lowestSignedProposalBytes := valBucket.Get(lowestSignedProposalKey)
+		if len(lowestSignedProposalBytes) == 0 {
+			return nil
+		}
+		lowestSignedProposalSlot = bytesutil.BytesToUint64BigEndian(lowestSignedProposalBytes)
+		return nil
+	})
+	return lowestSignedProposalSlot, err
+}
+
+// HighestSignedProposal returns the highest signed proposal slot for a validator public key.
+// If no data exists, returning 0 is a sensible default.
+func (store *Store) HighestSignedProposal(ctx context.Context, publicKey [48]byte) (uint64, error) {
+	ctx, span := trace.StartSpan(ctx, "Validator.HighestSignedProposal")
+	defer span.End()
+
+	var err error
+	var highestSignedProposalSlot uint64
+	err = store.view(func(tx *bolt.Tx) error {
+		bucket := tx.Bucket(newhistoricProposalsBucket)
+		valBucket := bucket.Bucket(publicKey[:])
+		if valBucket == nil {
+			return fmt.Errorf("validator history empty for public key: %#x", publicKey)
+		}
+		highestSignedProposalBytes := valBucket.Get(highestSignedProposalKey)
+		if len(highestSignedProposalBytes) == 0 {
+			return nil
+		}
+		highestSignedProposalSlot = bytesutil.BytesToUint64BigEndian(highestSignedProposalBytes)
+		return nil
+	})
+	return highestSignedProposalSlot, err
+}
+
 // UpdatePublicKeysBuckets for a specified list of keys.
-func (store *Store) OldUpdatePublicKeysBuckets(pubKeys [][48]byte) error {
+func (store *Store) UpdatePublicKeysBuckets(pubKeys [][48]byte) error {
 	return store.update(func(tx *bolt.Tx) error {
-		bucket := tx.Bucket(historicProposalsBucket)
+		bucket := tx.Bucket(newhistoricProposalsBucket)
 		for _, pubKey := range pubKeys {
 			if _, err := bucket.CreateBucketIfNotExists(pubKey[:]); err != nil {
 				return errors.Wrap(err, "failed to create proposal history bucket")
@@ -81,10 +171,12 @@ func (store *Store) OldUpdatePublicKeysBuckets(pubKeys [][48]byte) error {
 	})
 }
 
-func pruneProposalHistory(valBucket *bolt.Bucket, newestEpoch uint64) error {
+func pruneProposalHistoryBySlot(valBucket *bolt.Bucket, newestSlot uint64) error {
 	c := valBucket.Cursor()
 	for k, _ := c.First(); k != nil; k, _ = c.First() {
-		epoch := binary.LittleEndian.Uint64(k)
+		slot := bytesutil.BytesToUint64BigEndian(k)
+		epoch := helpers.SlotToEpoch(slot)
+		newestEpoch := helpers.SlotToEpoch(newestSlot)
 		// Only delete epochs that are older than the weak subjectivity period.
 		if epoch+params.BeaconConfig().WeakSubjectivityPeriod <= newestEpoch {
 			if err := c.Delete(); err != nil {
